chore(interop): cleanup supervisor client feat
